(fix) Refactoring to move each Injective data source class into its own module
